Collate component
Renamed the accumulator component to collate (to collect). Completed
the implementation of collate by using a internal observer array. We
need to move this to either the Component class or the Observer class.
Having a simple array of observers is good because we are not creating a
intermediate Observer object for each subscription. All observers are
called using setTimeout to avoid errors in one observer affecting
others.

A lot of TODOs left in the collate code.